@@ -172,31 +172,27 @@ public class HomeFragment extends BaseFragment implements MenuProvider {
             binding.apiVersion.setText(String.valueOf(ConfigManager.getXposedApiVersion()));
             binding.api.setText(ConfigManager.isDexObfuscateEnabled() ? R.string.enabled : R.string.not_enabled);
             binding.frameworkVersion.setText(String.format(LocaleDelegate.getDefaultLocale(), "%1$s (%2$d)", ConfigManager.getXposedVersionName(), ConfigManager.getXposedVersionCode()));
+            binding.managerPackageName.setText(activity.getPackageName());
             if (Build.VERSION.SDK_INT < Build.VERSION_CODES.Q) {
                 binding.dex2oatWrapper.setText(String.format(LocaleDelegate.getDefaultLocale(), "%s (%s)", getString(R.string.unsupported), getString(R.string.android_version_unsatisfied)));
             } else switch (ConfigManager.getDex2OatWrapperCompatibility()) {
-                case ILSPManagerService.DEX2OAT_OK:
-                    binding.dex2oatWrapper.setText(R.string.supported);
-                    break;
-                case ILSPManagerService.DEX2OAT_CRASHED:
-                    binding.dex2oatWrapper.setText(String.format(LocaleDelegate.getDefaultLocale(), "%s (%s)", getString(R.string.unsupported), getString(R.string.crashed)));
-                    break;
-                case ILSPManagerService.DEX2OAT_MOUNT_FAILED:
-                    binding.dex2oatWrapper.setText(String.format(LocaleDelegate.getDefaultLocale(), "%s (%s)", getString(R.string.unsupported), getString(R.string.mount_failed)));
-                    break;
-                case ILSPManagerService.DEX2OAT_SELINUX_PERMISSIVE:
-                    binding.dex2oatWrapper.setText(String.format(LocaleDelegate.getDefaultLocale(), "%s (%s)", getString(R.string.unsupported), getString(R.string.selinux_permissive)));
-                    break;
-                case ILSPManagerService.DEX2OAT_SEPOLICY_INCORRECT:
-                    binding.dex2oatWrapper.setText(String.format(LocaleDelegate.getDefaultLocale(), "%s (%s)", getString(R.string.unsupported), getString(R.string.sepolicy_incorrect)));
-                    break;
+                case ILSPManagerService.DEX2OAT_OK ->
+                        binding.dex2oatWrapper.setText(R.string.supported);
+                case ILSPManagerService.DEX2OAT_CRASHED ->
+                        binding.dex2oatWrapper.setText(String.format(LocaleDelegate.getDefaultLocale(), "%s (%s)", getString(R.string.unsupported), getString(R.string.crashed)));
+                case ILSPManagerService.DEX2OAT_MOUNT_FAILED ->
+                        binding.dex2oatWrapper.setText(String.format(LocaleDelegate.getDefaultLocale(), "%s (%s)", getString(R.string.unsupported), getString(R.string.mount_failed)));
+                case ILSPManagerService.DEX2OAT_SELINUX_PERMISSIVE ->
+                        binding.dex2oatWrapper.setText(String.format(LocaleDelegate.getDefaultLocale(), "%s (%s)", getString(R.string.unsupported), getString(R.string.selinux_permissive)));
+                case ILSPManagerService.DEX2OAT_SEPOLICY_INCORRECT ->
+                        binding.dex2oatWrapper.setText(String.format(LocaleDelegate.getDefaultLocale(), "%s (%s)", getString(R.string.unsupported), getString(R.string.sepolicy_incorrect)));
             }
         } else {
             binding.apiVersion.setText(R.string.not_installed);
             binding.api.setText(R.string.not_installed);
             binding.frameworkVersion.setText(R.string.not_installed);
+            binding.managerPackageName.setText(activity.getPackageName());
         }
-        binding.managerVersion.setText(String.format(LocaleDelegate.getDefaultLocale(), "%1$s (%2$d)", BuildConfig.VERSION_NAME, BuildConfig.VERSION_CODE));
 
         if (Build.VERSION.PREVIEW_SDK_INT != 0) {
             binding.systemVersion.setText(String.format(LocaleDelegate.getDefaultLocale(), "%1$s Preview (API %2$d)", Build.VERSION.CODENAME, Build.VERSION.SDK_INT));
@@ -222,9 +218,9 @@ public class HomeFragment extends BaseFragment implements MenuProvider {
                 "\n" +
                 binding.frameworkVersion.getText() +
                 "\n\n" +
-                activity.getString(R.string.info_manager_version) +
+                activity.getString(R.string.info_manager_package_name) +
                 "\n" +
-                binding.managerVersion.getText() +
+                binding.managerPackageName.getText() +
                 "\n\n" +
                 activity.getString(R.string.info_system_version) +
                 "\n" +
