@@ -23,7 +23,6 @@ package org.lsposed.manager.util;
 import android.content.Context;
 import android.content.pm.PackageInfo;
 import android.graphics.Bitmap;
-import android.graphics.drawable.AdaptiveIconDrawable;
 
 import androidx.annotation.NonNull;
 
@@ -44,10 +43,7 @@ public class AppModule extends AppGlideModule {
     @Override
     public void registerComponents(Context context, @NonNull Glide glide, Registry registry) {
         int iconSize = context.getResources().getDimensionPixelSize(R.dimen.app_icon_size);
-        var info = context.getApplicationInfo();
-        var pm = context.getPackageManager();
-        var shrinkNonAdaptiveIcons = info.loadIcon(pm) instanceof AdaptiveIconDrawable;
-        var factory = new AppIconModelLoader.Factory(iconSize, shrinkNonAdaptiveIcons, context);
+        var factory = new AppIconModelLoader.Factory(iconSize, false, context);
         registry.prepend(PackageInfo.class, Bitmap.class, factory);
     }
 }
