@@ -27,16 +27,19 @@ import androidx.annotation.NonNull;
 import org.lsposed.lspd.impl.LSPosedContext;
 import org.lsposed.lspd.util.Hookers;
 
-import de.robv.android.xposed.XC_MethodHook;
 import de.robv.android.xposed.XposedBridge;
 import de.robv.android.xposed.XposedInit;
 import de.robv.android.xposed.callbacks.XC_LoadPackage;
+import io.github.libxposed.api.XposedInterface;
 import io.github.libxposed.api.XposedModuleInterface;
+import io.github.libxposed.api.annotations.BeforeInvocation;
+import io.github.libxposed.api.annotations.XposedHooker;
 
-public class StartBootstrapServicesHooker extends XC_MethodHook {
+@XposedHooker
+public class StartBootstrapServicesHooker implements XposedInterface.Hooker {
 
-    @Override
-    protected void beforeHookedMethod(MethodHookParam<?> param) {
+    @BeforeInvocation
+    public static void beforeHookedMethod() {
         logD("SystemServer#startBootstrapServices() starts");
 
         try {
