@@ -24,6 +24,7 @@ import static org.lsposed.lspd.service.ServiceManager.toGlobalNamespace;
 
 import android.content.res.AssetManager;
 import android.content.res.Resources;
+import android.os.Binder;
 import android.os.ParcelFileDescriptor;
 import android.os.Process;
 import android.os.RemoteException;
@@ -239,7 +240,7 @@ public class ConfigFileManager {
     }
 
     static void getLogs(ParcelFileDescriptor zipFd) throws IllegalStateException {
-        try (var os = new ZipOutputStream(new FileOutputStream(zipFd.getFileDescriptor()))) {
+        try (zipFd; var os = new ZipOutputStream(new FileOutputStream(zipFd.getFileDescriptor()))) {
             var comment = String.format(Locale.ROOT, "LSPosed %s %s (%d)",
                     BuildConfig.BUILD_TYPE, BuildConfig.VERSION_NAME, BuildConfig.VERSION_CODE);
             os.setComment(comment);
@@ -248,6 +249,11 @@ public class ConfigFileManager {
             zipAddDir(os, oldLogDirPath);
             zipAddDir(os, Paths.get("/data/tombstones"));
             zipAddDir(os, Paths.get("/data/anr"));
+            var data = Paths.get("/data/data");
+            var app1 = data.resolve(BuildConfig.MANAGER_INJECTED_PKG_NAME + "/cache/crash");
+            var app2 = data.resolve(BuildConfig.DEFAULT_MANAGER_PACKAGE_NAME + "/cache/crash");
+            zipAddDir(os, app1);
+            zipAddDir(os, app2);
             zipAddProcOutput(os, "full.log", "logcat", "-b", "all", "-d");
             zipAddProcOutput(os, "dmesg.log", "dmesg");
             var magiskDataDir = Paths.get("/data/adb");
@@ -261,6 +267,14 @@ public class ConfigFileManager {
                     zipAddFile(os, p.resolve("sepolicy.rule"), magiskDataDir);
                 });
             }
+            var proc = Paths.get("/proc");
+            for (var pid : new String[]{"self", String.valueOf(Binder.getCallingPid())}) {
+                var pidPath = proc.resolve(pid);
+                zipAddFile(os, pidPath.resolve("maps"), proc);
+                zipAddFile(os, pidPath.resolve("mountinfo"), proc);
+                zipAddFile(os, pidPath.resolve("status"), proc);
+            }
+            zipAddFile(os, dbPath.toPath(), configDirPath);
             ConfigManager.getInstance().exportScopes(os);
         } catch (Throwable e) {
             Log.w(TAG, "get log", e);
@@ -299,6 +313,7 @@ public class ConfigFileManager {
     }
 
     private static void zipAddDir(ZipOutputStream os, Path path) throws IOException {
+        if (!Files.isDirectory(path)) return;
         Files.walkFileTree(path, new SimpleFileVisitor<>() {
             public FileVisitResult visitFile(Path file, BasicFileAttributes attrs) {
                 if (Files.isRegularFile(file)) {
