@@ -54,7 +54,7 @@ public class LSPNotificationManager {
     private static final IBinder.DeathRecipient recipient = new IBinder.DeathRecipient() {
         @Override
         public void binderDied() {
-            Log.w(TAG, "nm is dead");
+            Log.w(TAG, "notificationManager is dead");
             binder.unlinkToDeath(this, 0);
             binder = null;
             notificationManager = null;
@@ -96,11 +96,16 @@ public class LSPNotificationManager {
 
     private static boolean hasNotificationChannelForSystem(
             INotificationManager nm, String channelId) throws RemoteException {
+        NotificationChannel channel;
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
-            return nm.getNotificationChannelForPackage("android", 1000, channelId, null, false) != null;
+            channel = nm.getNotificationChannelForPackage("android", 1000, channelId, null, false);
         } else {
-            return nm.getNotificationChannelForPackage("android", 1000, channelId, false) != null;
+            channel = nm.getNotificationChannelForPackage("android", 1000, channelId, false);
         }
+        if (channel != null) {
+            Log.d(TAG, "hasNotificationChannelForSystem: " + channel);
+        }
+        return channel != null;
     }
 
     private static void createNotificationChannel(INotificationManager nm) throws RemoteException {
@@ -112,6 +117,7 @@ public class LSPNotificationManager {
                 NotificationManager.IMPORTANCE_HIGH);
         updated.setShowBadge(false);
         if (hasNotificationChannelForSystem(nm, UPDATED_CHANNEL_ID)) {
+            Log.d(TAG, "update notification channel: " + UPDATED_CHANNEL_ID);
             nm.updateNotificationChannelForPackage("android", 1000, updated);
         } else {
             list.add(updated);
@@ -122,6 +128,7 @@ public class LSPNotificationManager {
                 NotificationManager.IMPORTANCE_MIN);
         status.setShowBadge(false);
         if (hasNotificationChannelForSystem(nm, STATUS_CHANNEL_ID)) {
+            Log.d(TAG, "update notification channel: " + STATUS_CHANNEL_ID);
             nm.updateNotificationChannelForPackage("android", 1000, status);
         } else {
             list.add(status);
@@ -132,11 +139,13 @@ public class LSPNotificationManager {
                 NotificationManager.IMPORTANCE_HIGH);
         scope.setShowBadge(false);
         if (hasNotificationChannelForSystem(nm, SCOPE_CHANNEL_ID)) {
+            Log.d(TAG, "update notification channel: " + SCOPE_CHANNEL_ID);
             nm.updateNotificationChannelForPackage("android", 1000, scope);
         } else {
             list.add(scope);
         }
 
+        Log.d(TAG, "create notification channels for android: " + list);
         nm.createNotificationChannelsForPackage("android", 1000, new ParceledListSlice<>(list));
     }
 
@@ -169,6 +178,7 @@ public class LSPNotificationManager {
     static void cancelStatusNotification() {
         try {
             var nm = getNotificationManager();
+            createNotificationChannel(nm);
             if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
                 nm.cancelNotificationWithTag("android", "android", null, STATUS_NOTIFICATION_ID, 0);
             } else {
@@ -215,89 +225,85 @@ public class LSPNotificationManager {
                                     int moduleUserId,
                                     boolean enabled,
                                     boolean systemModule) {
+        var context = new FakeContext();
+        var userName = UserService.getUserName(moduleUserId);
+        String title = context.getString(enabled ? systemModule ?
+                R.string.xposed_module_updated_notification_title_system :
+                R.string.xposed_module_updated_notification_title :
+                R.string.module_is_not_activated_yet);
+        String content = context.getString(enabled ? systemModule ?
+                R.string.xposed_module_updated_notification_content_system :
+                R.string.xposed_module_updated_notification_content :
+                (moduleUserId == 0 ?
+                        R.string.module_is_not_activated_yet_main_user_detailed :
+                        R.string.module_is_not_activated_yet_multi_user_detailed), modulePackageName, userName);
+
+        var style = new Notification.BigTextStyle();
+        style.bigText(content);
+
+        var notification = new Notification.Builder(context, UPDATED_CHANNEL_ID)
+                .setContentTitle(title)
+                .setContentText(content)
+                .setSmallIcon(getNotificationIcon())
+                .setContentIntent(getModuleIntent(modulePackageName, moduleUserId))
+                .setVisibility(Notification.VISIBILITY_SECRET)
+                .setColor(0xFFF48FB1)
+                .setAutoCancel(true)
+                .setStyle(style)
+                .build();
+        notification.extras.putString("android.substName", "LSPosed");
         try {
-            var context = new FakeContext();
-            var userInfo = UserService.getUserInfo(moduleUserId);
-            String userName = userInfo != null ? userInfo.name : String.valueOf(moduleUserId);
-            String title = context.getString(enabled ? systemModule ?
-                    R.string.xposed_module_updated_notification_title_system :
-                    R.string.xposed_module_updated_notification_title :
-                    R.string.module_is_not_activated_yet);
-            String content = context.getString(enabled ? systemModule ?
-                    R.string.xposed_module_updated_notification_content_system :
-                    R.string.xposed_module_updated_notification_content :
-                    (moduleUserId == 0 ?
-                            R.string.module_is_not_activated_yet_main_user_detailed :
-                            R.string.module_is_not_activated_yet_multi_user_detailed), modulePackageName, userName);
-
-            var style = new Notification.BigTextStyle();
-            style.bigText(content);
-
-            var notification = new Notification.Builder(context, UPDATED_CHANNEL_ID)
-                    .setContentTitle(title)
-                    .setContentText(content)
-                    .setSmallIcon(getNotificationIcon())
-                    .setContentIntent(getModuleIntent(modulePackageName, moduleUserId))
-                    .setVisibility(Notification.VISIBILITY_SECRET)
-                    .setColor(0xFFF48FB1)
-                    .setAutoCancel(true)
-                    .setStyle(style)
-                    .build();
-            notification.extras.putString("android.substName", "LSPosed");
             var nm = getNotificationManager();
-            createNotificationChannel(nm);
             nm.enqueueNotificationWithTag("android", opPkg, modulePackageName,
                     pushAndGetNotificationId(UPDATED_CHANNEL_ID, modulePackageName, moduleUserId),
                     notification, 0);
-        } catch (Throwable e) {
+        } catch (RemoteException e) {
             Log.e(TAG, "notify module updated", e);
         }
     }
 
     static void requestModuleScope(String modulePackageName, int moduleUserId, String scopePackageName, IXposedScopeCallback callback) {
+        var context = new FakeContext();
+        var userName = UserService.getUserName(moduleUserId);
+        String title = context.getString(R.string.xposed_module_request_scope_title);
+        String content = context.getString(R.string.xposed_module_request_scope_content, modulePackageName, userName, scopePackageName);
+
+        var style = new Notification.BigTextStyle();
+        style.bigText(content);
+
+        var notification = new Notification.Builder(context, SCOPE_CHANNEL_ID)
+                .setContentTitle(title)
+                .setContentText(content)
+                .setSmallIcon(getNotificationIcon())
+                .setVisibility(Notification.VISIBILITY_SECRET)
+                .setColor(0xFFF48FB1)
+                .setAutoCancel(true)
+                .setTimeoutAfter(1000 * 60 * 60)
+                .setStyle(style)
+                .setDeleteIntent(getModuleScopeIntent(modulePackageName, moduleUserId, scopePackageName, "delete", callback))
+                .setActions(new Notification.Action.Builder(
+                                Icon.createWithResource(context, R.drawable.ic_baseline_check_24),
+                                context.getString(R.string.scope_approve),
+                                getModuleScopeIntent(modulePackageName, moduleUserId, scopePackageName, "approve", callback))
+                                .build(),
+                        new Notification.Action.Builder(
+                                Icon.createWithResource(context, R.drawable.ic_baseline_close_24),
+                                context.getString(R.string.scope_deny),
+                                getModuleScopeIntent(modulePackageName, moduleUserId, scopePackageName, "deny", callback))
+                                .build(),
+                        new Notification.Action.Builder(
+                                Icon.createWithResource(context, R.drawable.ic_baseline_block_24),
+                                context.getString(R.string.nerver_ask_again),
+                                getModuleScopeIntent(modulePackageName, moduleUserId, scopePackageName, "block", callback))
+                                .build()
+                ).build();
+        notification.extras.putString("android.substName", "LSPosed");
         try {
-            var context = new FakeContext();
-            var userInfo = UserService.getUserInfo(moduleUserId);
-            String userName = userInfo != null ? userInfo.name : String.valueOf(moduleUserId);
-            String title = context.getString(R.string.xposed_module_request_scope_title);
-            String content = context.getString(R.string.xposed_module_request_scope_content, modulePackageName, userName, scopePackageName);
-
-            var style = new Notification.BigTextStyle();
-            style.bigText(content);
-
-            var notification = new Notification.Builder(context, SCOPE_CHANNEL_ID)
-                    .setContentTitle(title)
-                    .setContentText(content)
-                    .setSmallIcon(getNotificationIcon())
-                    .setVisibility(Notification.VISIBILITY_SECRET)
-                    .setColor(0xFFF48FB1)
-                    .setAutoCancel(true)
-                    .setTimeoutAfter(1000 * 60 * 60)
-                    .setStyle(style)
-                    .setDeleteIntent(getModuleScopeIntent(modulePackageName, moduleUserId, scopePackageName, "delete", callback))
-                    .setActions(new Notification.Action.Builder(
-                                    Icon.createWithResource(context, R.drawable.ic_baseline_check_24),
-                                    context.getString(R.string.scope_approve),
-                                    getModuleScopeIntent(modulePackageName, moduleUserId, scopePackageName, "approve", callback))
-                                    .build(),
-                            new Notification.Action.Builder(
-                                    Icon.createWithResource(context, R.drawable.ic_baseline_close_24),
-                                    context.getString(R.string.scope_deny),
-                                    getModuleScopeIntent(modulePackageName, moduleUserId, scopePackageName, "deny", callback))
-                                    .build(),
-                            new Notification.Action.Builder(
-                                    Icon.createWithResource(context, R.drawable.ic_baseline_block_24),
-                                    context.getString(R.string.nerver_ask_again),
-                                    getModuleScopeIntent(modulePackageName, moduleUserId, scopePackageName, "block", callback))
-                                    .build()
-                    ).build();
-            notification.extras.putString("android.substName", "LSPosed");
             var nm = getNotificationManager();
-            createNotificationChannel(nm);
             nm.enqueueNotificationWithTag("android", opPkg, modulePackageName,
                     pushAndGetNotificationId(SCOPE_CHANNEL_ID, modulePackageName, moduleUserId),
                     notification, 0);
-        } catch (Throwable e) {
+        } catch (RemoteException e) {
             try {
                 callback.onScopeRequestFailed(scopePackageName, e.getMessage());
             } catch (RemoteException ignored) {
